refactor(be): DeliveryRepository 수정, SettlementSummaryProjection 수정 (#254)

+++ backend/src/main/java/com/deliveranything/domain/settlement/dto/SummaryResponse.java
@@ -1,12 +1,13 @@
 package com.deliveranything.domain.settlement.dto;
 
 import com.deliveranything.domain.settlement.dto.projection.SettlementSummaryProjection;
+import java.math.BigDecimal;
 
 public record SummaryResponse(
     Long totalTransactionCount,
     Long weeklyTransactionCount,
-    Long weeklySettledAmount,
-    Long monthlySettledAmount,
+    BigDecimal weeklySettledAmount,
+    BigDecimal monthlySettledAmount,
     Long totalSettledAmount,
     Long scheduledSettleAmount
 ) {
